updated client on about page
