+++ components/common/article-serie-card/index.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import Link from "next/link";
 import React, { useTransition } from "react";
 import cx from "classnames";
