+++ components/pages/home/blog-card/index.tsx
@@ -31,7 +31,6 @@ export default function HomeBlogCard({
       onMouseLeave={() => {
         blob.current?.classList.add("hidden");
       }}
-      aria-label={`Read article about ${title}`}
       ref={container}
       onMouseMove={handleMouseMove}
       href={`/articles/${slug}`}
