+++ hooks/useMouse.tsx
@@ -1,6 +1,5 @@
-import { PrimitiveProps } from "@react-three/fiber";
-import { useEffect, MutableRefObject } from "react";
-// ! WIP
+import { RefObject, useEffect } from "react";
+
 function getMouseDegrees(x: number, y: number, degreeLimit: number) {
   let dx = 0,
     dy = 0,
@@ -11,31 +10,22 @@ function getMouseDegrees(x: number, y: number, degreeLimit: number) {
 
   let w = { x: window.innerWidth, y: window.innerHeight };
 
-  // Left (Rotates neck left between 0 and -degreeLimit)
-  // 1. If cursor is in the left half of screen
   if (x <= w.x / 2) {
-    // 2. Get the difference between middle of screen and cursor position
     xdiff = w.x / 2 - x;
-    // 3. Find the percentage of that difference (percentage toward edge of screen)
     xPercentage = (xdiff / (w.x / 2)) * 100;
-    // 4. Convert that to a percentage of the maximum rotation we allow for the neck
     dx = ((degreeLimit * xPercentage) / 100) * -1;
   }
 
-  // Right (Rotates neck right between 0 and degreeLimit)
   if (x >= w.x / 2) {
     xdiff = x - w.x / 2;
     xPercentage = (xdiff / (w.x / 2)) * 100;
     dx = (degreeLimit * xPercentage) / 100;
   }
-  // Up (Rotates neck up between 0 and -degreeLimit)
   if (y <= w.y / 2) {
     ydiff = w.y / 2 - y;
     yPercentage = (ydiff / (w.y / 2)) * 100;
-    // Note that I cut degreeLimit in half when she looks up
     dy = ((degreeLimit * 0.5 * yPercentage) / 100) * -1;
   }
-  // Down (Rotates neck down between 0 and degreeLimit)
   if (y >= w.y / 2) {
     ydiff = y - w.y / 2;
     yPercentage = (ydiff / (w.y / 2)) * 100;
@@ -44,16 +34,7 @@ function getMouseDegrees(x: number, y: number, degreeLimit: number) {
   return { x: dx, y: dy };
 }
 
-export const useObjectMovement = (
-  object: MutableRefObject<PrimitiveProps>,
-  // primaryLight: MutableRefObject<DirectionalLight>,
-  // accentLight: MutableRefObject<DirectionalLight>
-) => {
-  // const [mousePosition, setMousePosition] = useState<{ x: number; y: number }>({
-  //   x: 0,
-  //   y: 0,
-  // });
-
+export const useObjectMovement = (object: RefObject<any>) => {
   const rotateModel = (evt: MouseEvent) => {
     const { x, y } = {
       x: 0,
@@ -68,16 +49,9 @@ export const useObjectMovement = (
   };
 
   useEffect(() => {
-    // if (typeof window !== "undefined") {
-    //   document.querySelector("#main").addEventListener("scroll", onScroll);
-    // }
-
     window.addEventListener("mousemove", rotateModel);
-    // window.addEventListener("wheel", onScroll);
     return () => {
       window.removeEventListener("mousemove", rotateModel);
-      // window.removeEventListener("wheel", onScroll);
-      //   document.querySelector("#main")?.removeEventListener("scroll", onScroll);
     };
   }, []);
   return { rotateModel };
