--- next.config.js
@@ -1,29 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "**.arkar.space",
-        port: "",
-      },
-      {
-        protocol: "https",
-        hostname: "cdn.sanity.io",
-        port: "",
-      },
-      {
-        protocol: "https",
-        hostname: "burma.social",
-        port: "",
-      },
-      {
-        protocol: "https",
-        hostname: "i.ytimg.com",
-        port: "",
-      },
-    ],
-  },
-};
-
-module.exports = nextConfig;
