+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "test": "jest --watch",
@@ -11,18 +11,9 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@algolia/autocomplete-core": "^1.17.0",
-    "@algolia/autocomplete-plugin-query-suggestions": "^1.17.0",
-    "@algolia/autocomplete-preset-algolia": "^1.17.0",
     "@heroicons/react": "^2.1.1",
-    "@liveblocks/client": "^1.9.8",
-    "@liveblocks/react": "^1.9.8",
     "@portabletext/react": "^2.0.3",
     "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-icons": "^1.3.0",
-    "@radix-ui/react-popover": "^1.0.7",
-    "@radix-ui/react-scroll-area": "^1.2.0",
-    "@radix-ui/react-slot": "^1.1.0",
     "@react-three/drei": "^9.115.0",
     "@react-three/fiber": "9.0.0-alpha.8",
     "@sanity/asset-utils": "^1.3.0",
@@ -39,13 +30,10 @@
     "@types/node": "^20.11.19",
     "@types/react": "^18.2.57",
     "@types/react-dom": "^18.2.19",
-    "@types/react-instantsearch-dom": "^6.12.7",
     "@types/three": "^0.149.0",
     "@vercel/analytics": "^0.1.11",
     "@vercel/og": "^0.0.27",
     "@vercel/postgres": "^0.3.2",
-    "algoliasearch": "^4.22.1",
-    "chart.js": "^4.4.1",
     "class-variance-authority": "^0.7.0",
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
@@ -53,7 +41,7 @@
     "dateformat": "^5.0.3",
     "dracula-prism": "^2.1.16",
     "eslint": "8.32.0",
-    "eslint-config-next": "14.2.3",
+    "eslint-config-next": "15.0.2",
     "form-data": "^4.0.0",
     "friendly-challenge": "^0.9.14",
     "html-to-text": "^9.0.5",
@@ -61,17 +49,15 @@
     "mailgun.js": "^8.2.2",
     "mastodon-api": "^1.3.0",
     "net": "^1.0.2",
-    "next": "15.0.1",
-    "next-sanity": "^9.0.14",
+    "next": "15.0.2",
+    "next-sanity": "^9.8.8",
     "next-sanity-image": "^6.1.1",
     "octokit": "^2.1.0",
     "posthog-js": "^1.147.0",
     "prismjs": "^1.29.0",
-    "react": "18.3.1",
+    "react": "19.0.0-rc-02c0e824-20241028",
     "react-cmdk": "^1.3.9",
-    "react-dom": "18.3.1",
-    "react-hotkeys-hook": "^4.5.0",
-    "react-instantsearch-dom": "^6.40.4",
+    "react-dom": "19.0.0-rc-02c0e824-20241028",
     "react-katex": "^3.0.1",
     "react-text-transition": "^3.1.0",
     "sanity": "3.40.0",
@@ -79,7 +65,6 @@
     "sanity-plugin-graph-view": "^2.0.1",
     "sanity-plugin-latex-input": "^2.0.4",
     "sharp": "^0.33.2",
-    "styled-components": "^6.1.9",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
     "three": "^0.169.0"
@@ -96,8 +81,8 @@
     "@types/jest": "^29.5.12",
     "@types/node": "18.11.18",
     "@types/prismjs": "^1.26.3",
-    "@types/react": "18.0.38",
-    "@types/react-dom": "18.0.10",
+    "@types/react": "npm:types-react@19.0.0-rc.1",
+    "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
     "@types/react-instantsearch-dom": "^6.12.3",
     "@types/testing-library__jest-dom": "^5.14.9",
     "@types/three": "^0.149.0",
@@ -106,7 +91,14 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.4.35",
+    "styled-components": "^6.1.13",
     "tailwindcss": "^3.4.1",
     "typescript": "^5.3.3"
+  },
+  "pnpm": {
+    "overrides": {
+      "@types/react": "npm:types-react@19.0.0-rc.1",
+      "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1"
+    }
   }
 }
