--- tailwind.config.js
@@ -1,83 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  darkMode: ["class"],
-  content: [
-    "./app/**/*.{js,ts,jsx,tsx}",
-    "./pages/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {
-      gridTemplateColumns: {
-        dashboard: "200px auto 300px",
-      },
-      gridTemplateRows: {
-        dashboard: "1fr",
-      },
-      colors: {
-        "theme-bg-secondary": "#16142d",
-        "theme-bg": "#131127",
-        "theme-primary": "#f056c8",
-        "theme-primary-opaque": "#f056c71a",
-        "theme-accent": "#88ced3",
-        "theme-accent-opaque": "#88ced309",
-        "theme-body": "#cbd5e1",
-
-        background: "hsl(var(--background))",
-        foreground: "hsl(var(--foreground))",
-
-        card: {
-          DEFAULT: "hsl(var(--card))",
-          foreground: "hsl(var(--card-foreground))",
-        },
-        popover: {
-          DEFAULT: "hsl(var(--popover))",
-          foreground: "hsl(var(--popover-foreground))",
-        },
-        primary: {
-          DEFAULT: "hsl(var(--primary))",
-          foreground: "hsl(var(--primary-foreground))",
-        },
-        secondary: {
-          DEFAULT: "hsl(var(--secondary))",
-          foreground: "hsl(var(--secondary-foreground))",
-        },
-        muted: {
-          DEFAULT: "hsl(var(--muted))",
-          foreground: "hsl(var(--muted-foreground))",
-        },
-        accent: {
-          DEFAULT: "hsl(var(--accent))",
-          foreground: "hsl(var(--accent-foreground))",
-        },
-        destructive: {
-          DEFAULT: "hsl(var(--destructive))",
-          foreground: "hsl(var(--destructive-foreground))",
-        },
-        border: "hsl(var(--border))",
-        input: "hsl(var(--input))",
-        ring: "hsl(var(--ring))",
-        chart: {
-          1: "hsl(var(--chart-1))",
-          2: "hsl(var(--chart-2))",
-          3: "hsl(var(--chart-3))",
-          4: "hsl(var(--chart-4))",
-          5: "hsl(var(--chart-5))",
-        },
-      },
-      fontFamily: {
-        theme: ["Ubuntu", "Inter", "sans-serif"],
-        design: ["Merriweather", "Inter", "sans-serif"],
-      },
-      borderRadius: {
-        lg: "var(--radius)",
-        md: "calc(var(--radius) - 2px)",
-        sm: "calc(var(--radius) - 4px)",
-      },
-    },
-  },
-  plugins: [
-    require("@tailwindcss/container-queries"),
-    require("tailwindcss-animate"),
-  ],
-};
