+++ types/index.ts
@@ -1,10 +1,10 @@
 export * as article from "./articles";
 
 export type DetailPageParamTypes = {
-  params?: {
-    slug?: string;
-  };
-  searchParams?: {
+  params: Promise<{
+    slug: string;
+  }>;
+  searchParams: Promise<{
     search?: string;
-  };
+  }>;
 };
