+++ utils/client.ts
@@ -1,9 +1,9 @@
 // client.js
-import {createClient}from "@sanity/client";
+import { createClient } from "@sanity/client";
 
 export default createClient({
   projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
   dataset: process.env.NEXT_PUBLIC_SANITY_DATASET,
-  apiVersion: "2021-10-21",
+  apiVersion: "vX",
   useCdn: true,
 });
